# Modulo_10/pond1/static/swagger.yaml
openapi: 3.0.0
info:
  title: TODO list
  description: To-do list por API!
  version: "1.0"
servers:
  - url: http://localhost:5000/
paths:
  /getAllTasks:
    get:
      summary: Lista todas as tarefas
      operationId: listTasks
      tags:
        - Tarefas
      responses:
        '200':
          description: Lista de tarefas retornada com sucesso
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Task'
      security:
        - basicAuth: []
  /createTask:
    post:
      summary: Cria uma nova tarefa e retorna
      operationId: createTask
      tags:
        - Tarefas
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskCreate'
      responses:
        '201':
          description: Tarefa criada com sucesso
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
      security:
        - basicAuth: []
  /getTask/{id}:
    get:
      summary: Obtém detalhes de uma tarefa específica por ID
      operationId: getTask
      tags:
        - Tarefas
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Detalhes da tarefa
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Task'
        '404':
          description: Tarefa não encontrada
      security:
        - basicAuth: []
  /updateTask/{id}:
    put:
      summary: Atualiza uma tarefa existente por ID
      operationId: updateTask
      tags:
        - Tarefas
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TaskUpdate'
      responses:
        '200':
          description: Tarefa atualizada com sucesso
      security:
        - basicAuth: []
  /deleteTask/{id}:
    delete:
      summary: Deleta uma tarefa por ID
      operationId: deleteTask
      tags:
        - Tarefas
      parameters:
        - name: task_id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Tarefa deletada com sucesso
        '404':
          description: Tarefa não encontrada
      security:
        - basicAuth: []
components:
  securitySchemes:
    basicAuth:
      type: http
      scheme: basic
  schemas:
    Task:
      type: object
      properties:
        id:
          type: integer
          example: 1
        title:
          type: string
          example: Terminar a documentação do Swagger
    TaskCreate:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          example: Nova Tarefa
    TaskUpdate:
      type: object
      properties:
        title:
          type: string
          example: Tarefa Atualizada